refactor(server): split handlers.rs into domain modules
Move 25+ HTTP handlers from the monolithic handlers.rs into eight
focused files under src/server/handlers/: agents, attachments,
channels, knowledge, messages, tasks, workspace, and mod (core
infrastructure + whoami/server-info). Zero behaviour changes.
